+++ sentry-samples/sentry-samples-spring/src/main/java/io/sentry/samples/spring/SentryConfig.java
@@ -1,5 +1,6 @@
 package io.sentry.samples.spring;
 
+import io.sentry.SentryOptions;
 import io.sentry.SentryOptions.TracesSamplerCallback;
 import io.sentry.spring.EnableSentry;
 import io.sentry.spring.tracing.SentryTracingConfiguration;
@@ -11,7 +12,8 @@ import org.springframework.context.annotation.Import;
 // project/dashboard
 @EnableSentry(
     dsn = "https://502f25099c204a2fbf4cb16edc5975d1@o447951.ingest.sentry.io/5428563",
-    sendDefaultPii = true)
+    sendDefaultPii = true,
+    maxRequestBodySize = SentryOptions.RequestSize.MEDIUM)
 @Import(SentryTracingConfiguration.class)
 public class SentryConfig {