+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 * Fix: set min sdk version of sentry-android-fragment to API 14 (#1608)
+* Feat: Add request body extraction for Spring MVC integration (#1595)
 
 ## 5.1.0-beta.5